fix: make sure to create picker before calling update - calling `.update()` before the date picker is created was causing the input to be override as the new picker element and was also causing leakage when clicking the input because at that point 2 elements were now assigned as the picker container (input & new picker in the body)
