+++ package/src/scripts/update.ts
@@ -1,5 +1,6 @@
 import { IReset } from '@src/types';
 import { VanillaCalendar } from '@src/vanilla-calendar';
+import { createCalendarToInput } from '@scripts/helpers/createCalendarToInput';
 import messages from '@scripts/helpers/getMessages';
 import reset from '@scripts/reset';
 
@@ -12,6 +13,9 @@ const update = (self: VanillaCalendar, {
 }: IReset = {}) => {
 	if (!self.isInit) throw new Error(messages.notInit);
 
+	if (self.input && !self.isInputInit) {
+		createCalendarToInput(self, false);
+	}
 	reset(self, {
 		year,
 		month,
