+++ package/src/types.ts
@@ -248,4 +248,5 @@ export interface IVanillaCalendar {
 	readonly dateMin: Date;
 	readonly dateMax: Date;
 	readonly isInit: boolean;
+	readonly isInputInit: boolean;
 }
